requirements.txt
flask
numpy
torch
Pillow
torchvision
requests
opencv-python